pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kb"
version = "0.1.0"
description = "Enterprise knowledge-base search — ingestion, retrieval, and generation"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Acme Platform" }]
license = { text = "MIT" }

dependencies = [
    # Config + types
    "pydantic>=2.8,<3",
    "pydantic-settings>=2.5,<3",
    "python-dotenv==1.0.1",

    # HTTP + API clients (pinned to what smoke tests validated)
    "requests==2.32.3",
    "httpx==0.27.2",
    "openai==1.55.2",

    # Datastores
    "psycopg[binary]==3.2.3",
    "redis==5.2.0",

    # Parsing
    "pypdf>=5.0",
    "pymupdf4llm>=0.0.17",
    "PyYAML>=6.0.1",

    # Tokenization (chunk sizing)
    "tiktoken>=0.8.0",

    # CLI
    "click>=8.1",

    # Observability
    "langsmith==0.1.147",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.7",
    "mypy>=1.11",
]

[project.scripts]
kb = "kb.cli:main"

[tool.setuptools.packages.find]
include = ["kb*"]
exclude = ["tests*", "scripts*", "data*", "docker*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501"]   # handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-ra --strict-markers"
